Fix table in "Data format (en).md"
The markdown formatting of the "Properties" example table in the English data format documentation was broken, at least on GitHub. I took the opportunity to fully align the table columns in the markdown source, but that is not necessary, as you only need to add a pipe at the start of each row which starts with an empty cell.